refactor: Move storage related functions to workload_storage.go
Storage related functions were initially placed in workload_single_node.go. This commit is extracting such storage related functions to its own workload_storage.go file. This is for better code organization in preparation to apply the same mechanism to more complex deployments (P/D, multi-workers, etc).
